+++ hivemind/client/optim/collaborative.py
@@ -18,10 +18,9 @@ LRSchedulerBase = getattr(torch.optim.lr_scheduler, '_LRScheduler', None)
 
 #TODO make sure that the weighted averager will NOT advertize for averaging if it is out of sync!
 # warn if expiration is too infrequent (measure mean time between steps)
-#TODO do not forget to setup client mode!
 #TODO make sure we can recover from zero_grad!
-#TODO actually use scheduler normally on step!
 #TODO enforce minimum period on reporting progress
+
 
 @dataclass(frozen=False)
 class CollaborationState:
